src/main/java/org/hobsoft/hamcrest/submatcher/SpyHolder.java
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 *     http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.hobsoft.hamcrest.submatcher;

/**
 * Holder for the current spy.
 */
final class SpyHolder
{
	// ----------------------------------------------------------------------------------------------------------------
	// fields
	// ----------------------------------------------------------------------------------------------------------------

	private static Spy<?> spy;
	
	// ----------------------------------------------------------------------------------------------------------------
	// constructors
	// ----------------------------------------------------------------------------------------------------------------

	private SpyHolder()
	{
		throw new AssertionError();
	}

	// ----------------------------------------------------------------------------------------------------------------
	// public methods
	// ----------------------------------------------------------------------------------------------------------------

	public static boolean hasSpy()
	{
		return (spy != null);
	}

	public static Spy<?> getSpy()
	{
		if (!hasSpy())
		{
			throw new IllegalStateException("Spy has not been set");
		}
		
		return spy;
	}

	public static void setSpy(Spy<?> spy)
	{
		SpyHolder.spy = spy;
	}
}
